% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available}
\alias{available}
\title{Discover projects with pre-computed HCA files}
\usage{
available(type = c("loom", "mtx.zip"))
}
\arguments{
\item{type}{`character(1)` type of archive to import. One of
`"loom"` (default) or `"mtx.zip"`.}
}
\value{
A `tibble` describing available projects and the full path
    to the archive.
}
\description{
Discover projects with pre-computed HCA files
}
\details{
This function visits
    \url{https://data.humancellatlas.org/explore/projects} and
    identifies projects for which the 'Matrix' file icon indicates
    that a matrix is available for download. The return value is a
    `tibble` that can be subset to a single project, and then the
    data retrieved to R using `import_loom()` or `import_mtxzip()`.
}
\examples{
dd <- available("mtx.zip")
dd
dd \%>\% select(projectTitle)
dd \%>\%
    filter(grepl("^A single-cell reference", projectTitle))
path <- dd \%>\%
    filter(row_number() == which.min(size)) \%>\%
    pull(path)
sce <- import_mtxzip(path)

}
