% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colData.R
\name{colDataTibble}
\alias{colDataTibble}
\alias{colDataConstants}
\alias{colDataBrief}
\title{Helper functions for simplifying colData}
\usage{
colDataTibble(sce)

colDataConstants(sce)

colDataBrief(sce)
}
\arguments{
\item{sce}{a `SummarizedExperiment` or derived object, e.g., a
`SingleCellExperiment`.}
}
\value{
`colDataTibble()` returns a tibble representation of
    `colData()`.

`colDataConstants()` returns a two-column tibble. The first
    column, `key` corresponds to `colData()` column names for which
    only a single distinct value is present. The second column
    `value` is the unique value.

`colDataBrief()` returns a tibble containing only those
    `colData()` columns with more than one value, and with column
    names abbreviated to the shortests common 'word' (using `_` or
    `.` as separators) suffixes.
}
\description{
`colDataTibble()` creates a tibble from `colData()`

`colDataConstants()` scans `colData()` columns to
    identify those with a single distinct value.

`colDataBrief()` produces a subset of `colData()`
    columns which contain more than one distinct value.
}
