% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R
\name{discover}
\alias{discover}
\title{Discover AWS S3 buckets with pre-computed HCA files}
\usage{
discover()
}
\value{
A `tibble` describing available samples and the full path
    to the archive.
}
\description{
Discover AWS S3 buckets with pre-computed HCA files
}
