% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_utilities.R
\name{replica}
\alias{replica}
\title{Useful functions for interacting with Human Cell Atlas resources}
\usage{
replica(x = c("aws", "gcp"))
}
\arguments{
\item{x}{character(1) replica to use, either `"aws"` or `"gcp"`.}
}
\value{
character(1) the (updated) replica
}
\description{
`replica()` sets or returns the replica for data retrieval.
}
\examples{
replica()

}
