% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mtx_zip.R
\name{import.mtxzip}
\alias{import.mtxzip}
\title{Import Human Cell Atlas '.mtx.zip' archives to SingleCellExperiment}
\usage{
import.mtxzip(
  path,
  ...,
  exdir = tempfile(),
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{path}{character(1) path to remote (`http://` or `https://`)
or local `.zip` archive.}

\item{...}{additional arguments, not supported.}

\item{exdir}{character(1) directory in which to extract .zip archive.}

\item{overwrite}{logical(1) overwrite existing files in `exdir`?}

\item{verbose}{logical(1) report progress using `message()`.}
}
\value{
`SingleCellExperiment()` representing the expreession
    data. Rows represent features and columns cells; counts are
    represented in a sparse matrix.
}
\description{
Import Human Cell Atlas '.mtx.zip' archives to SingleCellExperiment
}
\examples{
\dontrun{
url <- paste0(
    "https://s3.amazonaws.com/project-assets.data.humancellatlas.org/",
    "project-assets/project-matrices/",
    "116965f3-f094-4769-9d28-ae675c1b569c.homo_sapiens.mtx.zip"
)
sce <- import.mtxzip(url)
sce
}
}
