% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_utilities.R
\name{colTibble}
\alias{colTibble}
\alias{rowTibble}
\alias{constant}
\alias{brief}
\title{Helper functions for simplifying colData}
\usage{
colTibble(sce, ...)

rowTibble(sce, ...)

constant(.data)

brief(.data, abbreviate_names = TRUE)
}
\arguments{
\item{sce}{a `SummarizedExperiment` or derived object, e.g., a
`SingleCellExperiment`.}

\item{...}{additional arguments passed to `as_tibble()`, e.g., to
map rownames to a column.}

\item{.data}{A `data.frame()` or `tibble()`.}

\item{abbreviate_names}{logical(1) when `TRUE`, abbreviate column
names to the shortests common 'word' (using `[[:punct:]]+` as
separators) suffixes. E.g., column names `"a.b.c"` and `d.e.c`
would abbreviate to `b.c` and `e.c`.}
}
\value{
`colTibble()` returns a tibble representation of
    `colData()`.

`rowTibble() returns a tibble representation of
    `rowData()`.

`constant()` returns a two-column tibble. The first column,
    `key` corresponds to `colData()` column names for which only a
    single distinct value is present. The second column `value` is
    the unique value.

`brief()` returns a tibble containing only those
    `colData()` columns with more than one value, and with (by
    default) column names abbreviated to the shortests common
    'word' (using `[[:punct:]]+` as separators) suffixes.
}
\description{
`colTibble()` creates a tibble from `colData()`

`constant()` scans `colData()` columns to
    identify those with a single distinct value.

`brief()` produces a subset of `colData()`
    columns which contain more than one distinct value.
}
