% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colData_helpers.R
\name{colData_common}
\alias{colData_common}
\alias{colData_distinct}
\title{Helper functions for simpilfying colData}
\usage{
colData_common(sce)

colData_distinct(sce)
}
\value{
`colData_column()` returns a two-column tibble. The first
    column, `key` corresponds to `colData()` column names for which
    only a single distinct value is present. The second column
    `value` is the unique value.

`colData_distinct()` returns a tibble containing only those
    `colData()` columns with more than one value.
}
\description{
`colData_column()` scans `colData()` columns to
    identify those with a single distinct value.

`colData_distinct()` produces a subset of `colData()`
    columns which contain more than one distinct value.
}
