% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover.R
\name{discover}
\alias{discover}
\title{Discover projects with pre-computed HCA files}
\usage{
discover()
}
\value{
A `tibble` describing available projects and the full path
    to the archive.
}
\description{
Discover projects with pre-computed HCA files
}
\examples{
dd <- discover()
dd
dd \%>\% select(projectTitle)
dd \%>\%
    filter(grepl("^A single-cell reference", projectTitle)) \%>\%
    t()
path <- dd \%>\%
    filter(row_number() == which.min(size)) \%>\%
    pull(path)
sce <- import.mtxzip(path)

}
