% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_mtx_zip.R
\name{import.mtxzip}
\alias{import.mtxzip}
\title{Import Human Cell Atlas '.mtx.zip' archives to SingleCellExperiment}
\usage{
import.mtxzip(
  .data,
  path,
  ...,
  exdir = tempfile(),
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{.data}{(optional) When present, a data.frame or derived
    class, e.g., tibble, containing a single row, and a column
    specified by the `path` argument and containing a path to a
    remote (`http://` or `https://`) or local `.zip` archive.

    `.data` can also be a `character(1)` argument used to provide
    the path to the `.zip` archive directly.}

\item{path}{character(1) the path to the remote (`http://` or
`https://`) or local `.zip` archive or, when `.data` is
present, the column name in which the path is found.}

\item{...}{additional arguments, not supported.}

\item{exdir}{character(1) directory in which to extract .zip archive.}

\item{overwrite}{logical(1) overwrite existing files in `exdir`?}

\item{verbose}{logical(1) report progress using `message()`.}
}
\value{
`SingleCellExperiment()` representing the expreession
    data. Rows represent features and columns cells; counts are
    represented in a sparse matrix.
}
\description{
Import Human Cell Atlas '.mtx.zip' archives to SingleCellExperiment
}
\examples{
\dontrun{
aa <- available()
sce <-
    filter(aa, size == min(size)) %>%
    import.mtxzip()
}
}
